refactor(neon_rich_text): Separate building rich object from text span rendering
