+++ packages/neon_framework/packages/neon_rich_text/lib/src/rich_text.dart
@@ -3,10 +3,7 @@ import 'package:built_value/json_object.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 import 'package:intersperse/intersperse.dart';
-import 'package:neon_rich_text/src/rich_objects/deck_card.dart';
-import 'package:neon_rich_text/src/rich_objects/fallback.dart';
-import 'package:neon_rich_text/src/rich_objects/file.dart';
-import 'package:neon_rich_text/src/rich_objects/mention.dart';
+import 'package:neon_rich_text/src/rich_objects/rich_objects.dart';
 import 'package:nextcloud/core.dart' as core;
 
 /// Renders the [text] as a rich [TextSpan].
@@ -125,44 +122,5 @@ TextSpan buildRichTextSpan({
   return TextSpan(
     style: style,
     children: children,
-  );
-}
-
-/// Renders a rich object [parameter] to be interactive.
-InlineSpan buildRichObjectParameter({
-  required core.RichObjectParameter parameter,
-  required TextStyle? textStyle,
-  required bool isPreview,
-}) {
-  if (isPreview) {
-    return TextSpan(
-      text: parameter.name,
-      style: textStyle,
-    );
-  }
-
-  return WidgetSpan(
-    alignment: PlaceholderAlignment.middle,
-    child: switch (parameter.type) {
-      core.RichObjectParameter_Type.user ||
-      core.RichObjectParameter_Type.call ||
-      core.RichObjectParameter_Type.guest ||
-      core.RichObjectParameter_Type.userGroup =>
-        NeonRichObjectMention(
-          parameter: parameter,
-          textStyle: textStyle,
-        ),
-      core.RichObjectParameter_Type.file => NeonRichObjectFile(
-          parameter: parameter,
-          textStyle: textStyle,
-        ),
-      core.RichObjectParameter_Type.deckCard => NeonRichObjectDeckCard(
-          parameter: parameter,
-        ),
-      _ => NeonRichObjectFallback(
-          parameter: parameter,
-          textStyle: textStyle,
-        ),
-    },
   );
 }
