+++ packages/neon_framework/packages/neon_rich_text/pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
     sdk: flutter
   flutter_material_design_icons: ^1.1.7296
   intersperse: ^2.0.0
+  meta: ^1.0.0
   neon_framework:
     git:
       url: https://github.com/nextcloud/neon
